[brevtjeneste] Lagt til avtaleId på innsendingsmodell

+++ brevtjeneste/src/main/kotlin/no/david/brevtjeneste/model/Avtale.kt
@@ -3,8 +3,8 @@ package no.david.brevtjeneste.model
 import java.time.ZonedDateTime
 
 data class Avtale(
-    val id: Int,
-    val kundeId: Int,
+    val id: Long,
+    val kundeId: Long,
     val createdAt: ZonedDateTime,
     var status: AvtaleStatus
 )
